agent/requirements.txt
# Watchtower Agent - Local Dependencies

# Screen capture
mss>=9.0.0
Pillow>=10.0.0
imagehash>=4.3.0

# Input control
pyautogui>=0.9.54

# WebSocket client
websockets>=12.0

# System tray
pystray>=0.19.0

# PyInstaller (for building .exe)
pyinstaller>=6.0.0
